requirements.txt
streamlit
python-dotenv
accelerate
torch         
tiktoken
pydantic
jsonschema